pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai4bmr-learn"
version = "0.1.0"
description = "Minimal callbacks package for the pretrain-v6 workflow"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Adriano Martinelli", email = "adrianom@student.ethz.ch" }]

dependencies = [
  "colorcet",
  "geopandas",
  "glom",
  "loguru",
  "numpy",
  "opencv-python",
  "openslide-bin",
  "openslide-python",
  "pandas",
  "pillow",
  "pyarrow",
  "scikit-learn",
  "torch",
  "torchinfo",
  "tqdm",
  "lightning",
]

# Optional dependency groups (extras)
[project.optional-dependencies]
dev = [
  "black",
  "pytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

# Auto-discover packages under src/
[tool.setuptools.packages.find]
where = ["src"]

# If you ship non-.py files via MANIFEST.in, keep this true
[tool.setuptools.dynamic]
# (leave empty unless you dynamically load version/metadata)

[tool.black]
line-length = 120
